Rename to Sordini.
Sordini. In Das Sloß, a Castle bureaucrat of Italian descent who is famed for being especially conscientious. When the letter from the Gemeindevorsteher, the village mayor, rejecting the order to appoint a land surveyor falls mistakenly into Sordini's hands, he doggedly pursues the case in an attempt to clarify the situation. Ironically, his conscientiousness does not give rise to clarity but instead instigates a bureaucratic snarl of major proportions, and this bureaucratic error is responsible for K. being falsely appointed as land surveyor.